requirements.txt
aiogram==3.3.0
aiohttp==3.9.5
python-dotenv==1.0.0
finnhub-python==2.4.16
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
